--- src/domains/user/models/storesData.js
@@ -1,158 +1,0 @@
-export const storesData = [
-  {
-    id: 0,
-    name: "Lanches da Deusa",
-    discount: "Até 50% OFF",
-    rating: "5,0",
-    category: "Lancheria",
-    opening_hours: "10:00 às 22:00",
-    distancia: "1,2 Km",
-    isFavorited: true,
-    imagePath: "/src/ui/assets/image.png",
-    logoPath: "/src/ui/assets/image2.png",
-  },
-  {
-    id: 1,
-    name: "Dragão Verde",
-    discount: "",
-    rating: "4,7",
-    category: "Lancheria",
-    opening_hours: "11:00 às 23:00",
-    distancia: "3,4 Km",
-    isFavorited: false,
-    imagePath: "/src/ui/assets/dragao_verde_img.png",
-    logoPath: "/src/ui/assets/dragao_verde_logo.png",
-  },
-  {
-    id: 2,
-    name: "Artesanto do Pastel",
-    discount: "Compre 1, leve 2",
-    rating: "4,9",
-    category: "Pastelaria",
-    opening_hours: "08:00 às 14:00",
-    distancia: "2,1 Km",
-    isFavorited: true,
-    imagePath: "/src/ui/assets/artesanato_pastel_img.png",
-    logoPath: "/src/ui/assets/artesanato_pastel_logo.png",
-  },
-  {
-    id: 3,
-    name: "Doce Encanto",
-    discount: "Até 30% OFF",
-    rating: "4,8",
-    category: "Confeitaria",
-    opening_hours: "09:00 às 20:00",
-    distancia: "2,8 Km",
-    isFavorited: true,
-    imagePath: "/path/to/store4-image.png",
-    logoPath: "/path/to/store4-logo.png",
-  },
-  {
-    id: 4,
-    name: "Bistrô Gourmet",
-    discount: "10% OFF",
-    rating: "4,6",
-    category: "Restaurante",
-    opening_hours: "12:00 às 23:00",
-    distancia: "4,0 Km",
-    isFavorited: false,
-    imagePath: "/path/to/store5-image.png",
-    logoPath: "/path/to/store5-logo.png",
-  },
-  {
-    id: 5,
-    name: "Sabores do Mundo",
-    discount: "15% OFF",
-    rating: "4,5",
-    category: "Restaurante Internacional",
-    opening_hours: "13:00 às 22:00",
-    distancia: "3,2 Km",
-    isFavorited: false,
-    imagePath: "/path/to/store6-image.png",
-    logoPath: "/path/to/store6-logo.png",
-  },
-  {
-    id: 6,
-    name: "Sabores do Mundo",
-    discount: "15% OFF",
-    rating: "4,5",
-    category: "Restaurante Internacional",
-    opening_hours: "13:00 às 22:00",
-    distancia: "3,2 Km",
-    isFavorited: false,
-    imagePath: "/path/to/store6-image.png",
-    logoPath: "/path/to/store6-logo.png",
-  },
-  {
-    id: 7,
-    name: "Frutos do Mar",
-    discount: "Até 25% OFF",
-    rating: "4,9",
-    category: "Peixaria",
-    opening_hours: "10:00 às 21:00",
-    distancia: "2,5 Km",
-    isFavorited: false,
-    imagePath: "/path/to/store7-image.png",
-    logoPath: "/path/to/store7-logo.png",
-  },
-  {
-    id: 8,
-    name: "Churrascaria do Chef",
-    discount: "20% OFF",
-    rating: "4,7",
-    category: "Churrascaria",
-    opening_hours: "11:00 às 23:00",
-    distancia: "3,4 Km",
-    isFavorited: false,
-    imagePath: "/path/to/store8-image.png",
-    logoPath: "/path/to/store8-logo.png",
-  },
-  {
-    id: 9,
-    name: "Panificadora Pão Quente",
-    discount: "Compre 1, leve 2",
-    rating: "4,9",
-    category: "Panificadora",
-    opening_hours: "06:00 às 20:00",
-    distancia: "2,1 Km",
-    isFavorited: false,
-    imagePath: "/path/to/store9-image.png",
-    logoPath: "/path/to/store9-logo.png",
-  },
-  {
-    id: 10,
-    name: "Empório Gourmet",
-    discount: "Até 50% OFF",
-    rating: "5,0",
-    category: "Empório",
-    opening_hours: "10:00 às 22:00",
-    distancia: "1,2 Km",
-    isFavorited: false,
-    imagePath: "/path/to/store10-image.png",
-    logoPath: "/path/to/store10-logo.png",
-  },
-  {
-    id: 11,
-    name: "Sushi House",
-    discount: "20% OFF",
-    rating: "4,7",
-    category: "Sushi Bar",
-    opening_hours: "11:00 às 23:00",
-    distancia: "3,4 Km",
-    isFavorited: false,
-    imagePath: "/path/to/store11-image.png",
-    logoPath: "/path/to/store11-logo.png",
-  },
-  {
-    id: 12,
-    name: "Café Colonial",
-    discount: "Compre 1, leve 2",
-    rating: "4,9",
-    category: "Cafeteria",
-    opening_hours: "08:00 às 14:00",
-    distancia: "2,1 Km",
-    isFavorited: false,
-    imagePath: "/path/to/store12-image.png",
-    logoPath: "/path/to/store12-logo.png",
-  },
-]
